LF-407-2: drm/imx/dcss: fix module insertion hang
In dcss_dtg_sync_set(), the pixel clock is cut off in order to change the rate.
However, if the vblank interrupt happens in the same time, it will freeze the
platform as it will access the DTG registers while the clock is off.

This patch will activate the vblank interrupt after the dcss_dtg_sync_set() is
called.

Also, while at it, make sure the DPR and DTG interrupts are masked off before
requesting the interrupt, so that the handlers are not called before CRTC is
enabled.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>