arm64: dts: ls1028a-rdb: Enable SGMII AN for the QSGMII switch ports
This enables monitoring of link status and AN. It should also physically
enable SGMII AN with the VSC8514 PHY, but in practice that is still
hardcoded as "on" in the PHY driver, at the moment. So since Felix
actually disables SGMII AN when this DT property is absent, this would
result in an in-band AN mismatch between the MAC and the PHY. So this
property is required for the moment for this MAC/PHY combination.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>