net: phy: vsc8514: enable in-band SGMII auto-negotiation setting
The default in-band AN setting for the VSC8514 PHY is not very reliable:
its out-of-reset state is with SerDes AN disabled, but certain boot
firmware (such as U-Boot) enables it during the boot process.

So its final state as seen by Linux depends on whether the U-Boot PHY
driver has run or not.

If SGMII auto-negotiation is enabled but not acknowledged by the MAC,
the PHY does not pass traffic.

But otherwise, it is able to pass traffic both with AN disabled, and
with AN enabled.

We would like to make this explicitly configurable rather than hardcoded
as "on" as we are doing right now, but we'd rather hardcode it in LSDK
and wait until a solution lands upstream, than invent a solution for
this here.

Signed-off-by: Alex Marginean <alexandru.marginean@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>